+++ scripts/add-covid-prep_category.sql
@@ -1,2 +1,0 @@
-alter table `application` CHANGE `prep_category` `prep_category`
-enum('cov','mic','rml','tgs','wes','wgs','wts') DEFAULT NULL;