+++ scripts/add-created_at-updated_at-to-family_sample.sql
@@ -1,5 +1,0 @@
-ALTER TABLE `family_sample`
-ADD COLUMN `created_at` datetime DEFAULT NULL;
-
-ALTER TABLE `family_sample`
-ADD COLUMN `updated_at` datetime DEFAULT NULL;