+++ scripts/remove_microbial_tables.sql
@@ -1,3 +1,0 @@
-DROP TABLE flowcell_microbial_sample;
-DROP TABLE microbial_sample;
-DROP TABLE microbial_order;